requirements.txt
discord.py
sqlalchemy
peewee
pymysql
psutil
python-dotenv